spectrum-27: refactor to hook and create feature folder structure
