+++ app/features/tasks/EditTask.tsx
@@ -1,25 +1,23 @@
-import { useEffect, useState } from "react";
-import { Button } from "../atoms/button";
-import { Input } from "../atoms/input";
-import { Label } from "../atoms/label";
+import { Button } from "../../components/atoms/button";
+import { Input } from "../../components/atoms/input";
+import { Label } from "../../components/atoms/label";
 import {
   Dialog,
   DialogContent,
   DialogFooter,
   DialogHeader,
   DialogTitle,
-  DialogTrigger,
-} from "../ui/dialog";
+} from "../../components/ui/dialog";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "../ui/select";
-import { Textarea } from "../ui/textarea";
-import { useActionData, useSubmit } from "@remix-run/react";
+} from "../../components/ui/select";
+import { Textarea } from "../../components/ui/textarea";
 import type { Task } from "@prisma/client";
+import { useEditTask } from "./hooks/useEditTask";
 
 interface EditTaskProps {
   open: boolean;
@@ -32,61 +30,20 @@ export default function EditTask({
   setOpen,
   selectedTask,
 }: EditTaskProps) {
-  const [title, setTitle] = useState(""); // TODO: [title, setTitle
-  const [description, setDescription] = useState("");
-  const [dueDate, setDueDate] = useState("");
-  const [priority, setPriority] = useState("");
-  const [disableButton, setDisableButton] = useState(true); // TODO: [disableButton, setDisableButton
-  const submit = useSubmit();
-  const actionData = useActionData();
+  const {
+    title,
+    setTitle,
+    description,
+    setDescription,
+    dueDate,
+    setDueDate,
+    priority,
+    setPriority,
+    disableButton,
+    handleSubmit,
+    actionData,
+  } = useEditTask({ open, setOpen, selectedTask });
 
-  useEffect(() => {
-    if (actionData?.success) {
-      setOpen(false);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [actionData]);
-
-  useEffect(() => {
-    if (selectedTask) {
-      setTitle(selectedTask.title);
-      setDescription(selectedTask.description);
-      if (selectedTask.dueDate) {
-        const date = new Date(selectedTask.dueDate);
-        // format date to yyyy-MM-dd
-        const formattedDate = date.toISOString().split("T")[0];
-        setDueDate(formattedDate);
-      }
-      setPriority(selectedTask.priority);
-    }
-  }, [selectedTask]);
-
-  useEffect(() => {
-    if (title && description && dueDate && priority) {
-      setDisableButton(false);
-    } else {
-      setDisableButton(true);
-    }
-  }, [title, description, dueDate, priority]);
-
-  const handleSubmit = () => {
-    // check make sure values from state are not empty
-    if (!title || !description || !dueDate || !priority) {
-      return;
-    }
-    const formData = new FormData();
-    formData.append("taskId", selectedTask.id.toString());
-    formData.append("title", title);
-    formData.append("description", description);
-    formData.append("dueDate", dueDate);
-    formData.append("priority", priority);
-    formData.append("intent", "edit");
-    submit(formData, {
-      method: "post",
-    });
-  };
-
-  console.log(dueDate);
   return (
     <Dialog open={open} onOpenChange={setOpen}>
       <DialogContent className="sm:max-w-[425px] bg-white">
